requirements.txt
Flask==3.1.3
gunicorn==23.0.0
cloudinary==1.42.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
pymongo==4.16.0
PyMySQL==1.1.2
python-dotenv==1.2.2
python-slugify==8.0.4
SQLAlchemy==2.0.49
Werkzeug==3.1.8
WTForms==3.2.1
